requirements.txt
yfinance
pandas
requests
beautifulsoup4
openai
requests
pyyaml